+++ sshmitm/interfaces/server.py
@@ -1,12 +1,12 @@
 import logging
 import os
 import struct
-from typing import TYPE_CHECKING, Any, List, Optional, Tuple, Type, Union
+from typing import TYPE_CHECKING, Any, cast
 
 import paramiko
 from paramiko.message import Message
 from paramiko.pkey import PKey
-from paramiko.sftp import _VERSION, CMD_INIT, CMD_VERSION, SFTPError
+from paramiko.sftp import CMD_INIT, CMD_VERSION, SFTPError
 
 from sshmitm.clients.netconf import NetconfClient
 from sshmitm.clients.sftp import SFTPClient
@@ -22,19 +22,20 @@ if TYPE_CHECKING:
     )
 
 
+_VERSION = 3
+
+
 class BaseServerInterface(paramiko.ServerInterface, BaseModule):
     def __init__(self, session: "sshmitm.session.Session") -> None:
         super().__init__()
-        self.session: "sshmitm.session.Session" = session
+        self.session: sshmitm.session.Session = session
         self.session.register_session_thread()
-        self.forwarders: List[
-            Union[
-                TunnelForwarder,
-                LocalPortForwardingForwarder,
-                RemotePortForwardingForwarder,
-            ]
+        self.forwarders: list[
+            TunnelForwarder
+            | LocalPortForwardingForwarder
+            | RemotePortForwardingForwarder
         ] = []
-        self.possible_auth_methods: Optional[List[str]] = None
+        self.possible_auth_methods: list[str] | None = None
 
 
 class ServerInterface(BaseServerInterface):
@@ -251,8 +252,8 @@ class ServerInterface(BaseServerInterface):
         return paramiko.common.AUTH_FAILED
 
     def check_auth_interactive(
-        self, username: str, submethods: Union[bytes, str]
-    ) -> Union[int, paramiko.server.InteractiveQuery]:
+        self, username: str, submethods: bytes | str
+    ) -> int | paramiko.server.InteractiveQuery:
         logging.debug(
             "check_auth_interactive: username=%s, submethods=%s", username, submethods
         )
@@ -269,8 +270,8 @@ class ServerInterface(BaseServerInterface):
         return auth_interactive_query
 
     def check_auth_interactive_response(
-        self, responses: List[str]
-    ) -> Union[int, paramiko.server.InteractiveQuery]:
+        self, responses: list[str]
+    ) -> int | paramiko.server.InteractiveQuery:
         logging.debug("check_auth_interactive_response: responses=%s", responses)
         is_trivial_auth = (
             self.args.enable_trivial_auth and self.session.accepted_key is not None
@@ -416,7 +417,7 @@ class ServerInterface(BaseServerInterface):
         self.session.ssh_client.transport.cancel_port_forward(address, port)
 
     def check_channel_direct_tcpip_request(
-        self, chanid: int, origin: Tuple[str, int], destination: Tuple[str, int]
+        self, chanid: int, origin: tuple[str, int], destination: tuple[str, int]
     ) -> int:
         username = self.session.transport.get_username()
         logging.info(
@@ -467,7 +468,7 @@ class ServerInterface(BaseServerInterface):
         channel: paramiko.Channel,
         single_connection: bool,
         auth_protocol: str,
-        auth_cookie: Union[bytes, bytearray],
+        auth_cookie: bytes | bytearray,
         screen_number: int,
     ) -> bool:
         logging.debug(
@@ -482,7 +483,7 @@ class ServerInterface(BaseServerInterface):
 
     def check_global_request(
         self, kind: str, msg: paramiko.message.Message
-    ) -> Union[bool, Tuple[Union[bool, int, str], ...]]:
+    ) -> bool | tuple[bool | int | str, ...]:
         logging.debug("check_global_request: kind=%s, msg=%s", kind, msg)
         return False
 
@@ -493,7 +494,7 @@ class ProxySFTPServer(paramiko.SFTPServer):
         channel: paramiko.Channel,
         name: str,
         server: ServerInterface,
-        sftp_si: Type[paramiko.SFTPServerInterface],
+        sftp_si: type[paramiko.SFTPServerInterface],
         session: "sshmitm.session.Session",
         *largs: Any,
         **kwargs: Any,
@@ -510,14 +511,14 @@ class ProxySFTPServer(paramiko.SFTPServer):
         # original implementaion:
         # https://github.com/paramiko/paramiko/blob/d9ab89a0f8ae37a25d44565d5eb03a5d93fed5b9/paramiko/sftp.py#L153
 
-        t, data = self._read_packet()
+        t, data = self._read_packet()  # type: ignore[attr-defined]
         if t != CMD_INIT:
             raise SFTPError("Incompatible sftp protocol")  # noqa: TRY003,EM101
         version = struct.unpack(">I", data[:4])[0]
         msg = Message()
         msg.add_int(_VERSION)
-        self._send_packet(CMD_VERSION, msg)
-        return version
+        self._send_packet(CMD_VERSION, msg)  # type: ignore[attr-defined]
+        return cast("int", version)
 
     def start_subsystem(
         self, name: str, transport: paramiko.Transport, channel: paramiko.Channel