+++ sshmitm/moduleparser/utils.py
@@ -1,7 +1,8 @@
 import argparse
-from typing import TYPE_CHECKING, Any, Dict, Optional, Sequence, Type, Union
+from collections.abc import Sequence
+from typing import TYPE_CHECKING, Any
 
-from colored.colored import attr, fg  # type: ignore[import-untyped]
+from colored.colored import attr, fg
 
 from sshmitm.logger import Colors
 from sshmitm.utils import metadata
@@ -10,7 +11,7 @@ if TYPE_CHECKING:
     from sshmitm.moduleparser.modules import BaseModule
 
 
-def load_module(entry_point_class: Type["BaseModule"]) -> Type["argparse.Action"]:
+def load_module(entry_point_class: type["BaseModule"]) -> type["argparse.Action"]:
     """Action to be able to define BaseModule with the "add_module" method of the ModuleParser as command line parameter"""
 
     class ModuleLoaderAction(argparse.Action):
@@ -18,8 +19,8 @@ def load_module(entry_point_class: Type["BaseModule"]) -> Type["argparse.Action"
             self,
             parser: argparse.ArgumentParser,
             namespace: argparse.Namespace,
-            values: Union[str, Sequence[Any], None],
-            option_string: Optional[str] = None,
+            values: str | Sequence[Any] | None,
+            option_string: str | None = None,
         ) -> None:
             del parser
             del option_string
@@ -36,8 +37,8 @@ def load_module(entry_point_class: Type["BaseModule"]) -> Type["argparse.Action"
 
 
 def set_module_kwargs(
-    entry_point_class: Type["BaseModule"], **kwargs: Any
-) -> Dict[str, Any]:
+    entry_point_class: type["BaseModule"], **kwargs: Any
+) -> dict[str, Any]:
     entry_points = sorted(
         metadata.entry_points(group=f"sshmitm.{entry_point_class.__name__}"),
         key=lambda x: x.name,